MIDIPlayer now mutes/unmutes tracks based on wetness values
Currently it will create random sets of 40 wetness values then
mute/unmute each channel if the value is above an activation amount
(half way wet).  The next update is to have it actually read from the
serial like it will for the arduino.  We'll also need to build something
that will create the 40 values and write them to the serial like the
arduino will - like an arduino simulator.